+++ test/jasmine/spec/SpecHelper.js
@@ -171,7 +171,5 @@ var getDataAtCell = handsontableMethodFactory('getDataAtCell');
 var alter = handsontableMethodFactory('alter');
 var loadData = handsontableMethodFactory('loadData');
 var destroyEditor = handsontableMethodFactory('destroyEditor');
-var setCellReadOnly = handsontableMethodFactory('setCellReadOnly');
-var setCellEditable = handsontableMethodFactory('setCellEditable');
 var render = handsontableMethodFactory('render');
 var updateSettings = handsontableMethodFactory('updateSettings');